GFORMS-2685 - Allow users to return to CYA page (change status to InProgress) regardless of destinations
