+++ app/uk/gov/hmrc/gform/gform/AcknowledgementController.scala
@@ -203,35 +203,4 @@ class AcknowledgementController(
       Future.successful(Redirect(s"/feedback/${formTemplateId.value}").withNewSession)
     }
 
-  def changeStateAndRedirectToCYA(
-    formTemplateId: FormTemplateId,
-    maybeAccessCode: Option[AccessCode]
-  ): Action[AnyContent] =
-    auth.authAndRetrieveForm[SectionSelectorType.Normal](
-      formTemplateId,
-      maybeAccessCode,
-      OperationWithForm.ForceUpdateFormStatus
-    ) { implicit request => _ => cache => _ => _ =>
-      for {
-        _ <- gformConnector.deleteGeneratedFiles(cache.form.envelopeId)
-        res <-
-          gformConnector
-            .updateUserData(
-              FormIdData(cache.retrievals, formTemplateId, maybeAccessCode),
-              UserData(
-                cache.form.formData,
-                InProgress,
-                cache.form.visitsIndex,
-                cache.form.thirdPartyData,
-                cache.form.componentIdToFileId
-              )
-            )
-            .flatMap { _ =>
-              Future.successful(
-                Redirect(routes.SummaryController.summaryById(formTemplateId, maybeAccessCode, None, None, true, None))
-              )
-            }
-      } yield res
-    }
-
 }
