+++ conf/app.routes
@@ -44,7 +44,6 @@ GET           /declaration/:formTemplateId/:maybeAccessCode                     
 POST          /declaration/:formTemplateId/:maybeAccessCode                                                      uk.gov.hmrc.gform.gform.DeclarationController.submitDeclaration(formTemplateId: FormTemplateId, maybeAccessCode: Option[AccessCode], action: Direction)
 GET           /acknowledgement/:formTemplateId                                                                   uk.gov.hmrc.gform.gform.AcknowledgementController.showAcknowledgement(a: Option[AccessCode], formTemplateId: FormTemplateId)
 GET           /acknowledgement/pdf/:formTemplateId                                                               uk.gov.hmrc.gform.gform.AcknowledgementController.downloadPDF(a: Option[AccessCode], formTemplateId: FormTemplateId)
-GET           /acknowledgement/return-to-cya/:formTemplateId                                                     uk.gov.hmrc.gform.gform.AcknowledgementController.changeStateAndRedirectToCYA(formTemplateId: FormTemplateId, a: Option[AccessCode])
 GET           /error/insufficient-enrolments/:formTemplateId                                                     uk.gov.hmrc.gform.auth.ErrorController.insufficientEnrolments(formTemplateId: FormTemplateId)
 GET           /error/browser-fordidden/:formTemplateId                                                           uk.gov.hmrc.gform.auth.ErrorController.browserForbidden(formTemplateId: FormTemplateId)
 GET           /error/file-limit-exceed/:formTemplateId                                                           uk.gov.hmrc.gform.gform.SummaryController.handleFileLimitExceeded(formTemplateId: FormTemplateId, a: Option[AccessCode], uploadedFileSize: Long)
